Switched to material components theme
Grouped fragments, activities, threads and utils to separate threads
Added tab layouts to dashboard and contract details
Added upcoming and expired installments to dashboard
