+++ app/src/main/java/app/agrivest/android/activities/DashboardActivity.java
@@ -1,24 +1,23 @@
-package app.agrivest.android;
+package app.agrivest.android.activities;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.app.AlertDialog;
-import android.app.ProgressDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TableLayout;
-import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.DefaultRetryPolicy;
@@ -29,27 +28,32 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.google.android.material.tabs.TabLayout;
 
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
-public class DashboardActivity extends AppCompatActivity implements View.OnClickListener {
+import app.agrivest.android.fragments.CommitmentsFragment;
+import app.agrivest.android.fragments.ContractsFragment;
+import app.agrivest.android.R;
+import app.agrivest.android.utils.SQLiteHelper;
+import app.agrivest.android.utils.Utils;
+import app.agrivest.android.api.API;
+
+public class DashboardActivity extends AppCompatActivity {
     private Utils utils;
     private boolean connected;
     private SharedPreferences userDetails;
     private RequestQueue mQueue;
 
-    private Button refresh_contracts_BT;
+    private Toolbar toolbar;
+    private ViewPager viewPager;
+    private TabLayout tabLayout;
 
-    private TextView network_status_message;
-    private TextView last_updated_message;
-
-    private TableLayout contracts_TL;
-
-    private ProgressDialog loadContractsDialog;
-    private ProgressDialog uploadOfflineReceiptsDialog;
+    private ContractsFragment contractsFragment;
+    private CommitmentsFragment commitmentsFragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -61,29 +65,55 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
         userDetails = this.getSharedPreferences("user_details", MODE_PRIVATE);
         mQueue = Volley.newRequestQueue(this);
 
-        refresh_contracts_BT = findViewById(R.id.refresh_contracts_BT);
-        refresh_contracts_BT.setOnClickListener(this);
+        toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
 
-        setNetworkStatusMessage();
-        lastUpdated();
-        loadContracts();
-        uploadOfflineReceipts();
+        viewPager = findViewById(R.id.view_pager);
+        tabLayout = findViewById(R.id.tab_layout);
+
+        contractsFragment = new ContractsFragment();
+        commitmentsFragment = new CommitmentsFragment();
+
+        tabLayout.setupWithViewPager(viewPager);
+
+        ViewPagerAdapter viewPagerAdapter = new ViewPagerAdapter(getSupportFragmentManager(), 0);
+        viewPagerAdapter.addFragment(contractsFragment, "Contracts");
+        viewPagerAdapter.addFragment(commitmentsFragment, "Commitments");
+        viewPager.setAdapter(viewPagerAdapter);
+        tabLayout.getTabAt(0).setIcon(R.drawable.ic_people_black_24dp);
+        tabLayout.getTabAt(1).setIcon(R.drawable.ic_comment_black_24dp);
     }
 
-    private void setNetworkStatusMessage() {
-        network_status_message = findViewById(R.id.network_status_message);
-        if (connected) {
-            network_status_message.setTextColor(Color.parseColor("#21634b"));
-            network_status_message.setText("Connected!");
-        } else {
-            network_status_message.setTextColor(Color.parseColor("#948b23"));
-            network_status_message.setText("Disconnected! You are seeing data that was cached when you were last connected");
+    private class ViewPagerAdapter extends FragmentPagerAdapter {
+
+        private List<Fragment> fragments = new ArrayList<>();
+        private List<String> fragmentTitle = new ArrayList<>();
+
+        public ViewPagerAdapter(@NonNull FragmentManager fm, int behavior) {
+            super(fm, behavior);
         }
-    }
 
-    private void lastUpdated() {
-        last_updated_message = findViewById(R.id.last_updated_message);
-        last_updated_message.setText("Last updated " + userDetails.getString("last_updated", ""));
+        public void addFragment(Fragment fragment, String title) {
+            fragments.add(fragment);
+            fragmentTitle.add(title);
+        }
+
+        @NonNull
+        @Override
+        public Fragment getItem(int position) {
+            return fragments.get(position);
+        }
+
+        @Override
+        public int getCount() {
+            return fragments.size();
+        }
+
+        @Nullable
+        @Override
+        public CharSequence getPageTitle(int position) {
+            return fragmentTitle.get(position);
+        }
     }
 
     @Override
@@ -97,8 +127,7 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
 
         connected = utils.isInternetAvailable(getApplicationContext());
 
-//        setNetworkStatusMessage();
-//        uploadOfflineReceipts();
+        uploadOfflineReceipts();
     }
 
     private void signOut() {
@@ -120,61 +149,6 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
         }
     }
 
-    private void loadContracts() {
-        loadContractsDialog = new ProgressDialog(this);
-
-        contracts_TL = findViewById(R.id.contracts_TL);
-        final SQLiteDatabase db = new SQLiteHelper(getApplicationContext()).getReadableDatabase();
-
-        while (contracts_TL.getChildCount() > 1)
-            contracts_TL.removeView(contracts_TL.getChildAt(contracts_TL.getChildCount() - 1));
-
-        connected = utils.isInternetAvailable(getApplicationContext());
-        if (connected) {
-            lastUpdated();
-            loadContractsDialog.setTitle("Loading Contracts From Server");
-            loadContractsDialog.setMessage("Please wait while contracts are loaded");
-            loadContractsDialog.setCancelable(false);
-            if (!isFinishing()) {
-                loadContractsDialog.show();
-            }
-            String url = new API().getApiLink() + "/contract/search?search=&state=&officer=" + userDetails.getString("id", "") + "&batch=";
-            StringRequest request = new StringRequest(Request.Method.GET, url, new Response.Listener<String>() {
-                @Override
-                public void onResponse(String response) {
-                    SetResponseContractsThread setResponseContractsThread = new SetResponseContractsThread(getApplicationContext(), db, response, contracts_TL, loadContractsDialog);
-                    SharedPreferences.Editor userEditor = userDetails.edit();
-                    userEditor.putString("last_updated", new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(Calendar.getInstance().getTime()));
-                    userEditor.commit();
-                    lastUpdated();
-                }
-            }, new Response.ErrorListener() {
-                @Override
-                public void onErrorResponse(VolleyError error) {
-                    loadContractsDialog.dismiss();
-                    error.printStackTrace();
-                    signOut();
-                }
-            }) {
-                @Override
-                public Map<String, String> getHeaders() throws AuthFailureError {
-                    Map<String, String> params = new HashMap<String, String>();
-                    params.put("Authorization", "Bearer " + userDetails.getString("token", ""));
-                    params.put("Content-Type", "application/x-www-form-urlencoded");
-                    return params;
-                }
-            };
-            request.setRetryPolicy(new DefaultRetryPolicy(10000, 0, DefaultRetryPolicy.DEFAULT_BACKOFF_MULT));
-            mQueue.add(request);
-        } else {
-            loadContractsDialog.setTitle("Loading Contracts From Cache");
-            loadContractsDialog.setMessage("Please wait while contracts are loaded");
-            loadContractsDialog.setCancelable(false);
-            loadContractsDialog.show();
-            LoadLocalContractsThread loadLocalContractsThread = new LoadLocalContractsThread(this, db, contracts_TL, loadContractsDialog);
-        }
-    }
-
     public void uploadReceipt(final SQLiteDatabase db, final Cursor res) {
         String url = new API().getApiLink() + "/contract/receipt";
         StringRequest request = new StringRequest(Request.Method.POST, url, new Response.Listener<String>() {
@@ -182,14 +156,14 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
             public void onResponse(String response) {
                 db.delete(SQLiteHelper.RECEIPT_TABLE_NAME, "id = ?", new String[]{res.getString(res.getColumnIndex("id"))});
                 res.moveToNext();
-                if(res.isAfterLast() == false)
+                if (res.isAfterLast() == false)
                     uploadReceipt(db, res);
             }
         }, new Response.ErrorListener() {
             @Override
             public void onErrorResponse(VolleyError error) {
                 NetworkResponse networkResponse = error.networkResponse;
-                if(networkResponse != null) {
+                if (networkResponse != null) {
                     switch (networkResponse.statusCode) {
                         case 500:
                             Toast.makeText(getApplicationContext(), "Error uploading offline receipts. Please check the amounts and validate with the office", Toast.LENGTH_LONG).show();
@@ -201,7 +175,7 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
                     Log.d("STATUS_CODE", String.valueOf(networkResponse.statusCode));
                 }
                 res.moveToNext();
-                if(res.isAfterLast() == false)
+                if (res.isAfterLast() == false)
                     uploadReceipt(db, res);
                 error.printStackTrace();
             }
@@ -212,7 +186,7 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
                 params.put("cid", res.getString(res.getColumnIndex("contract_id")));
                 params.put("user_id", res.getString(res.getColumnIndex("user_id")));
                 params.put("amount", res.getString(res.getColumnIndex("amount")));
-                if(res.getString(res.getColumnIndex("payment_type")).equals("Cash")) {
+                if (res.getString(res.getColumnIndex("payment_type")).equals("Cash")) {
                     params.put("due_date", "");
                 } else {
                     params.put("due_date", res.getString(res.getColumnIndex("due_date")));
@@ -257,19 +231,5 @@ public class DashboardActivity extends AppCompatActivity implements View.OnClick
                 return super.onOptionsItemSelected(item);
         }
         return true;
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.refresh_contracts_BT:
-                connected = utils.isInternetAvailable(getApplicationContext());
-                if(!connected) {
-                    Toast.makeText(getApplicationContext(), "You are offline", Toast.LENGTH_SHORT).show();
-                } else {
-                    loadContracts();
-                }
-                break;
-        }
     }
 }
