+++ app/src/main/java/app/agrivest/android/api/API.java
@@ -1,4 +1,4 @@
-package app.agrivest.android;
+package app.agrivest.android.api;
 
 public class API {
     private final String APILINK = "https://www.agrivest.app/api-dev";
