+++ app/src/main/java/app/agrivest/android/fragments/ReceiptsFragment.java
@@ -1,14 +1,15 @@
-package app.agrivest.android;
+package app.agrivest.android.fragments;
 
-import androidx.appcompat.app.AppCompatActivity;
 
 import android.app.ProgressDialog;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.Button;
+import android.view.ViewGroup;
 import android.widget.TableLayout;
+
+import androidx.fragment.app.Fragment;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
@@ -22,17 +23,22 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.text.SimpleDateFormat;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class IssuedReceiptsActivity extends AppCompatActivity implements View.OnClickListener {
+import app.agrivest.android.R;
+import app.agrivest.android.api.API;
+import app.agrivest.android.utils.RowAdder;
+import app.agrivest.android.utils.Utils;
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class ReceiptsFragment extends Fragment {
 
     Utils utils;
 
@@ -41,56 +47,55 @@ public class IssuedReceiptsActivity extends AppCompatActivity implements View.On
     private SharedPreferences userDetails;
     private RequestQueue mQueue;
 
-    JSONArray receipts = null;
-
     TableLayout receiptsTable;
-    Button print_summary_BT;
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_issued_receipts);
 
-        getSupportActionBar().setTitle("Issued Receipts");
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-        utils = new Utils();
-        userDetails = this.getSharedPreferences("user_details", MODE_PRIVATE);
-        mQueue = Volley.newRequestQueue(this);
-        receiptsTable = findViewById(R.id.receipts_TL);
-        print_summary_BT = findViewById(R.id.print_summary_BT);
-        print_summary_BT.setOnClickListener(this);
-        loadReceipts();
+    public ReceiptsFragment() {
+        // Required empty public constructor
     }
 
 
-    private void loadReceipts() {
-        if(utils.isInternetAvailable(this)) {
-            loadReceipts = new ProgressDialog(this);
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View receipts = inflater.inflate(R.layout.fragment_receipts, container, false);
+
+        final String id = getArguments().getString("id");
+
+        utils = new Utils();
+        userDetails = getActivity().getSharedPreferences("user_details", getActivity().MODE_PRIVATE);
+        mQueue = Volley.newRequestQueue(getActivity());
+        receiptsTable = receipts.findViewById(R.id.receipts_TL);
+        loadReceipts(id);
+
+        return receipts;
+    }
+
+    private void loadReceipts(String id) {
+        if(utils.isInternetAvailable(getActivity())) {
+            loadReceipts = new ProgressDialog(getActivity());
             loadReceipts.setTitle("Loading receipts");
             loadReceipts.setMessage("Please wait while receipts are loaded");
             loadReceipts.setCancelable(false);
-            loadReceipts.show();
-            SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-            String url = new API().getApiLink() + "/contract/receipts/officer/" + userDetails.getString("id", "") + "/" + formatter.format(new Date());
+            String url =  new API().getApiLink() + "/contract/receipts/" + id;
             StringRequest request = new StringRequest(Request.Method.GET, url, new Response.Listener<String>() {
                 @Override
                 public void onResponse(String response) {
                     try {
-                        receipts = new JSONArray(response);
-                        RowAdder rowAdder = new RowAdder(getApplicationContext());
-                        for (int i = 0; i < receipts.length(); i++) {
-                            JSONObject receipt = receipts.getJSONObject(i);
-                            Iterator<String> installmentIterator = receipt.keys();
+                        JSONArray installments = new JSONArray(response);
+                        RowAdder rowAdder = new RowAdder(getActivity());
+                        for (int i = 0; i < installments.length(); i++) {
+                            JSONObject installment = installments.getJSONObject(i);
+                            Iterator<String> installmentIterator = installment.keys();
                             String tableExp[] = {};
                             final LinkedHashMap<String, String> contractMap = new LinkedHashMap<String, String>();
                             while (installmentIterator.hasNext()) {
                                 String key = installmentIterator.next();
                                 if (!Arrays.asList(tableExp).contains(key)) {
                                     if(key.equals("notes")) {
-                                        contractMap.put(key, new JSONObject(receipt.getString(key)).getString("String"));
+                                        contractMap.put(key, new JSONObject(installment.getString(key)).getString("String"));
                                     } else {
-                                        contractMap.put(key, receipt.getString(key));
+                                        contractMap.put(key, installment.getString(key));
                                     }
                                 }
                             }
@@ -119,20 +124,6 @@ public class IssuedReceiptsActivity extends AppCompatActivity implements View.On
             };
 
             mQueue.add(request);
-        }
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch(v.getId()) {
-            case R.id.print_summary_BT:
-                loadReceipts = new ProgressDialog(this);
-                loadReceipts.setTitle("Printing receipt summary");
-                loadReceipts.setMessage("Please wait while receipt summary is printed");
-                loadReceipts.setCancelable(false);
-                loadReceipts.show();
-                PrintReceiptSummaryThread printReceiptSummaryThread = new PrintReceiptSummaryThread(this, loadReceipts, receipts);
-                break;
         }
     }
 }
