+++ app/src/main/java/app/agrivest/android/threads/IssueReceiptThread.java
@@ -1,6 +1,6 @@
-package app.agrivest.android;
+package app.agrivest.android.threads;
 
-import android.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
@@ -14,6 +14,7 @@ import android.graphics.BitmapFactory;
 import android.os.Handler;
 
 import com.android.volley.AuthFailureError;
+import com.android.volley.DefaultRetryPolicy;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
@@ -22,10 +23,6 @@ import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.bxl.config.editor.BXLConfigLoader;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.nio.ByteBuffer;
 import java.util.Date;
 import java.util.HashMap;
@@ -33,26 +30,32 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
+import app.agrivest.android.api.API;
+import app.agrivest.android.utils.NumberFormatter;
+import app.agrivest.android.R;
+import app.agrivest.android.utils.SQLiteHelper;
+import app.agrivest.android.utils.Utils;
 import jpos.JposConst;
 import jpos.JposException;
 import jpos.POSPrinter;
 import jpos.POSPrinterConst;
 import jpos.config.JposEntry;
 
-public class PrintReceiptSummaryThread implements Runnable {
+public class IssueReceiptThread implements Runnable {
     Thread thrd;
     Context context;
     ProgressDialog progressDialog;
-    JSONArray receipts;
+    HashMap<String, String> receiptDetails;
     Utils utils;
     AlertDialog printReceiptStatus;
     Handler mainHandler;
     SharedPreferences userDetails;
+    RequestQueue mQueue;
 
-    PrintReceiptSummaryThread(Context context, ProgressDialog progressDialog, JSONArray receipts) {
+    public IssueReceiptThread(Context context, ProgressDialog progressDialog, HashMap<String, String> receiptDetails) {
         this.context = context;
         this.progressDialog = progressDialog;
-        this.receipts = receipts;
+        this.receiptDetails = receiptDetails;
         utils = new Utils();
         thrd = new Thread(this);
         thrd.start();
@@ -83,16 +86,6 @@ public class PrintReceiptSummaryThread implements Runnable {
     @Override
     public void run() {
         mainHandler = new Handler(context.getMainLooper());
-
-        if(receipts == null || receipts.length() == 0) {
-            progressDialog.dismiss();
-            showMessage(
-                    "Failure",
-                    "No receipts issued",
-                    "You do not have any receipts issued today."
-            );
-            return;
-        }
 
         BluetoothAdapter bluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
         Set<BluetoothDevice> bondedDeviceSet = bluetoothAdapter.getBondedDevices();
@@ -159,7 +152,6 @@ public class PrintReceiptSummaryThread implements Runnable {
         }
 
         try {
-            userDetails = context.getSharedPreferences("user_details", context.MODE_PRIVATE);
             POSPrinter posPrinter = new POSPrinter(context);
             posPrinter.open(BXLConfigLoader.PRODUCT_NAME_SPP_R210);
             posPrinter.claim(5000);
@@ -172,53 +164,45 @@ public class PrintReceiptSummaryThread implements Runnable {
             buffer.put((byte) 0x00);
             Bitmap logo = BitmapFactory.decodeResource(context.getResources(), R.drawable.logo);
             Bitmap payment_confirmation_customer_information = BitmapFactory.decodeResource(context.getResources(), R.drawable.payment_confirmation_customer_information);
+
+            String address = "\nAgrivest (Private) Limited\n" +
+                    "Hospital Junction, Polonnaruwa\n " +
+                    "027 222 22279\n";
+
+            if (receiptDetails.get("agrivest").equals("0")) {
+                logo = BitmapFactory.decodeResource(context.getResources(), R.drawable.randeepa_logo);
+                address = "\nRandeepa Agrarian (Pvt) Limited\n" +
+                        "Sirisangabo Place, Polonnaruwa\n " +
+                        "027 222 22279\n";
+            }
+
             posPrinter.printBitmap(buffer.getInt(0), logo, 200, POSPrinterConst.PTR_BM_CENTER);
             String ESCAPE_CHARACTERS = new String(new byte[]{0x1b, 0x7c});
-            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "N" + ESCAPE_CHARACTERS + "cA" + "\nAgrivest (Private) Limited\n" +
-                    "Hospital Junction, Polonnaruwa\n " +
-                    "027 222 22279\n");
+            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "N" + ESCAPE_CHARACTERS + "cA" + address);
             NumberFormatter numberFormatter = new NumberFormatter();
             posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "lA" + ESCAPE_CHARACTERS + "N"
                     + "________________________________\n\n");
-            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "bC" + ESCAPE_CHARACTERS + "cA" + "RECEIPT SUMMARY" + "\n\n");
+            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "bC" + ESCAPE_CHARACTERS + "cA" + "OFFICIAL RECEIPT" + "\n\n");
             posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "lA" + ESCAPE_CHARACTERS + "N"
-                    + java.text.DateFormat.getDateTimeInstance().format(new Date()) + "\n" + userDetails.getString("name", "") + "\n\n");
-
-            try {
-                double total = 0;
-                for (int i = 0; i < receipts.length(); i++) {
-                    JSONObject receipt = receipts.getJSONObject(i);
-                    float amount = Float.parseFloat(receipt.getString("amount"));
-                    total += amount;
-                    posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "lA" + ESCAPE_CHARACTERS + "N"
-                            + receipt.getString("id") + "\tRS. " + numberFormatter.format(amount) + "\n");
-                }
-                posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "N" + ESCAPE_CHARACTERS + "cA"
-                        + "\nTotal\n");
-                posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "4C" + ESCAPE_CHARACTERS + "cA"
-                        + "RS. " + numberFormatter.format(total) + "\n\n\n");
-            } catch (JSONException e) {
-                e.printStackTrace();
-                posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, "\n\n\n");
-                posPrinter.close();
-                progressDialog.dismiss();
-                showMessage(
-                        "Failure",
-                        "Something went wrong",
-                        "Something went wrong."
-                );
-                return;
-            }
-
+                    + java.text.DateFormat.getDateTimeInstance().format(new Date()) + "\n"
+                    + "C/ID: " + receiptDetails.get("id") + "\n"
+                    + receiptDetails.get("customer_name") + " (" + receiptDetails.get("customer_contact") + ")" + "\n"
+                    + "Chassis: " + receiptDetails.get("chassis_number") + "\n"
+                    + "Total Payable: RS. "
+                    + numberFormatter.format(receiptDetails.get("total_payable")) + "\n"
+                    + "Arrears: RS. "
+                    + numberFormatter.format(receiptDetails.get("amount_pending")) + "\n\n");
+            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "N" + ESCAPE_CHARACTERS + "cA" + "Recieved sum of" + "\n");
+            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "4C" + ESCAPE_CHARACTERS + "cA" + "RS " + numberFormatter.format(receiptDetails.get("amount")) + "\n");
+            posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, ESCAPE_CHARACTERS + "N" + ESCAPE_CHARACTERS + "cA" + "Thank You!" + "\n\n");
+            ByteBuffer buffer2 = ByteBuffer.allocate(4);
+            buffer2.put((byte) POSPrinterConst.PTR_S_RECEIPT);
+            buffer2.put((byte) 20);
+            buffer2.put((byte) 0x01);
+            buffer2.put((byte) 0x00);
+            posPrinter.printBitmap(buffer2.getInt(0), payment_confirmation_customer_information, 384, POSPrinterConst.PTR_BM_CENTER);
             posPrinter.printNormal(POSPrinterConst.PTR_S_RECEIPT, "\n\n\n");
             posPrinter.close();
-            progressDialog.dismiss();
-            showMessage(
-                    "success",
-                    "Printed",
-                    "Receipt summary printed."
-            );
-            return;
         } catch (JposException e) {
             e.printStackTrace();
             showMessage(
@@ -230,6 +214,81 @@ public class PrintReceiptSummaryThread implements Runnable {
                             "3. Paper roll not empty\n" +
                             "4. Paper cover closed"
             );
+            return;
         }
+
+        userDetails = context.getSharedPreferences("user_details", context.MODE_PRIVATE);
+        if (utils.isInternetAvailable(context)) {
+            mQueue = Volley.newRequestQueue(context);
+            String url = new API().getApiLink() + "/contract/receipt";
+            StringRequest request = new StringRequest(Request.Method.POST, url, new Response.Listener<String>() {
+                @Override
+                public void onResponse(String response) {
+                    showMessage(
+                            "success",
+                            "[ONLINE] Receipt issued",
+                            "Receipt number is " + response
+                    );
+                }
+            }, new Response.ErrorListener() {
+                @Override
+                public void onErrorResponse(VolleyError error) {
+                    saveReceipt();
+                    error.printStackTrace();
+                }
+            }) {
+                @Override
+                protected Map<String, String> getParams() {
+                    Map<String, String> params = new HashMap<String, String>();
+                    params.put("cid", receiptDetails.get("id"));
+                    params.put("user_id", userDetails.getString("id", ""));
+                    params.put("amount", receiptDetails.get("amount"));
+                    if (receiptDetails.get("payment_type").equals("Cash")) {
+                        params.put("due_date", "");
+                    } else {
+                        params.put("due_date", receiptDetails.get("due_date"));
+                    }
+                    return params;
+                }
+
+                @Override
+                public Map<String, String> getHeaders() throws AuthFailureError {
+                    Map<String, String> params = new HashMap<String, String>();
+                    params.put("Authorization", "Bearer " + userDetails.getString("token", ""));
+                    params.put("Content-Type", "application/x-www-form-urlencoded");
+                    return params;
+                }
+            };
+            request.setRetryPolicy(new DefaultRetryPolicy(10000, 0, DefaultRetryPolicy.DEFAULT_BACKOFF_MULT));
+            mQueue.add(request);
+        } else {
+            saveReceipt();
+        }
+    }
+
+    private void saveReceipt() {
+        SQLiteDatabase db = new SQLiteHelper(context).getReadableDatabase();
+        ContentValues receiptValues = new ContentValues();
+        receiptValues.put(SQLiteHelper.RECEIPT_CONTRACT_ID, receiptDetails.get("id"));
+        receiptValues.put(SQLiteHelper.RECEIPT_CUSTOMER_NAME, receiptDetails.get("customer_name"));
+        receiptValues.put(SQLiteHelper.RECEIPT_USER_ID, userDetails.getString("id", ""));
+        receiptValues.put(SQLiteHelper.RECEIPT_AMOUNT, receiptDetails.get("amount"));
+        receiptValues.put(SQLiteHelper.RECEIPT_PAYMENT_TYPE, receiptDetails.get("payment_type"));
+        receiptValues.put(SQLiteHelper.RECEIPT_DUE_DATE, receiptDetails.get("due_date"));
+        long id = db.insert(SQLiteHelper.RECEIPT_TABLE_NAME, null, receiptValues);
+        if (id != -1) {
+            showMessage(
+                    "success",
+                    "[OFFLINE] Receipt issued",
+                    "Please connect to internet as soon as possible."
+            );
+        } else {
+            showMessage(
+                    "failure",
+                    "[OFFLINE] Receipt failed",
+                    "Failed to saved offline receipt data. Please contact system administrator immediately."
+            );
+        }
+        db.close();
     }
 }
