+++ app/src/main/java/app/agrivest/android/utils/NumberFormatter.java
@@ -1,4 +1,4 @@
-package app.agrivest.android;
+package app.agrivest.android.utils;
 
 import java.text.DecimalFormat;
