+++ app/src/main/java/app/agrivest/android/utils/RowAdder.java
@@ -1,4 +1,4 @@
-package app.agrivest.android;
+package app.agrivest.android.utils;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,30 +6,34 @@ import android.graphics.Color;
 import android.graphics.Typeface;
 import android.net.Uri;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TableLayout;
 import android.widget.TableRow;
 import android.widget.TextView;
 
-import java.text.DecimalFormat;
+import com.google.android.material.button.MaterialButton;
+
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import app.agrivest.android.R;
+import app.agrivest.android.activities.ContractDetailsActivity;
+import app.agrivest.android.activities.ReceiptActivity;
+
 public class RowAdder {
     Context context;
-    RowAdder(Context context) {
+    public RowAdder(Context context) {
         this.context = context;
     }
     public void contract(TableLayout table, LinkedHashMap<String, String> contract) {
         TableRow contractRow = new TableRow(context);
-        contractRow.setBackgroundColor(Color.parseColor("#d5f0dc"));
+        contractRow.setBackgroundColor(Color.parseColor("#f2eded"));
         contractRow.setPadding(5, 5, 5, 5);
 
         Iterator contractIterator = contract.entrySet().iterator();
         while (contractIterator.hasNext()) {
             Map.Entry contractColumn = (Map.Entry) contractIterator.next();
-            Button btn = new Button(context);
+            MaterialButton btn = new MaterialButton(context, null, R.attr.materialButtonOutlinedStyle);
             TextView textView = new TextView(context);
             String columnValue = contractColumn.getValue().toString();
 
@@ -51,14 +55,6 @@ public class RowAdder {
 
             if (contractColumn.getKey().equals("id") || contractColumn.getKey().equals("total_payable")) {
                 final String id = contract.get("id");
-                final String agrivest = contract.get("agrivest");
-                final String customer_name = contract.get("customer_name");
-                final String customer_address = contract.get("customer_address");
-                final String model = contract.get("model");
-                final String customer_contact = contract.get("customer_contact");
-                final String chassis_number = contract.get("chassis_number");
-                final String amount_pending = contract.get("amount_pending");
-                final String total_payable = contract.get("total_payable");
 
                 btn.setClickable(true);
                 btn.setText(columnValue);
@@ -69,14 +65,6 @@ public class RowAdder {
                         Intent intent = (columnKey.equals("id")) ? new Intent(context, ContractDetailsActivity.class) : new Intent(context, ReceiptActivity.class);
                         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                         intent.putExtra("id", id);
-                        intent.putExtra("agrivest", agrivest);
-                        intent.putExtra("customer_name", customer_name);
-                        intent.putExtra("customer_address", customer_address);
-                        intent.putExtra("model", model);
-                        intent.putExtra("customer_contact", customer_contact);
-                        intent.putExtra("chassis_number", chassis_number);
-                        intent.putExtra("amount_pending", amount_pending);
-                        intent.putExtra("total_payable", total_payable);
                         context.startActivity(intent);
                     }
                 });
@@ -112,13 +100,12 @@ public class RowAdder {
 
     public void receipt(TableLayout table, LinkedHashMap<String, String> contract) {
         TableRow contractRow = new TableRow(context);
-        contractRow.setBackgroundColor(Color.parseColor("#d5f0dc"));
+        contractRow.setBackgroundColor(Color.parseColor("#f2eded"));
         contractRow.setPadding(5, 5, 5, 5);
 
         Iterator contractIterator = contract.entrySet().iterator();
         while (contractIterator.hasNext()) {
             Map.Entry contractColumn = (Map.Entry) contractIterator.next();
-            Button btn = new Button(context);
             TextView textView = new TextView(context);
             String columnValue = contractColumn.getValue().toString();
 
@@ -141,13 +128,12 @@ public class RowAdder {
 
     public void installment(TableLayout table, LinkedHashMap<String, String> contract) {
         TableRow contractRow = new TableRow(context);
-        contractRow.setBackgroundColor(Color.parseColor("#d5f0dc"));
+        contractRow.setBackgroundColor(Color.parseColor("#f2eded"));
         contractRow.setPadding(5, 5, 5, 5);
 
         Iterator contractIterator = contract.entrySet().iterator();
         while (contractIterator.hasNext()) {
             Map.Entry contractColumn = (Map.Entry) contractIterator.next();
-            Button btn = new Button(context);
             TextView textView = new TextView(context);
             String columnValue = contractColumn.getValue().toString();
 
@@ -171,6 +157,65 @@ public class RowAdder {
             textView.setText(columnValue);
             textView.setPadding(10, 10, 10, 10);
             contractRow.addView(textView);
+        }
+
+        table.addView(contractRow);
+    }
+
+    public void commitment(TableLayout table, LinkedHashMap<String, String> contract, String type) {
+        TableRow contractRow = new TableRow(context);
+        contractRow.setBackgroundColor(Color.parseColor("#f2eded"));
+        contractRow.setPadding(5, 5, 5, 5);
+
+        Iterator contractIterator = contract.entrySet().iterator();
+        while (contractIterator.hasNext()) {
+            Map.Entry contractColumn = (Map.Entry) contractIterator.next();
+            MaterialButton btn = new MaterialButton(context, null, R.attr.materialButtonOutlinedStyle);
+            TextView textView = new TextView(context);
+            String columnValue = contractColumn.getValue().toString();
+            if (contractColumn.getKey().equals("installment") || contractColumn.getKey().equals("installment_paid")) {
+                double amount = Double.parseDouble(contractColumn.getValue().toString());
+                NumberFormatter formatter = new NumberFormatter();
+
+                columnValue = formatter.format(amount);
+            }
+
+            textView.setText(columnValue);
+
+            if (contractColumn.getKey().equals("contract_id") || contractColumn.getKey().equals("total_payable")) {
+                final String id = contract.get("contract_id");
+
+                btn.setClickable(true);
+                btn.setText(columnValue);
+                final String columnKey = contractColumn.getKey().toString();
+                btn.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Intent intent = (columnKey.equals("contract_id")) ? new Intent(context, ContractDetailsActivity.class) : new Intent(context, ReceiptActivity.class);
+                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        intent.putExtra("id", id);
+                        context.startActivity(intent);
+                    }
+                });
+            }
+
+            if (contractColumn.getKey().equals("due_in")) {
+                if(type.equals("upcoming")) {
+                    textView.setText("Expires in " + Math.abs(Integer.parseInt(columnValue)) + " days");
+                    textView.setTextColor(Color.parseColor("#196912"));
+                } else {
+                    textView.setText("Expired " + Math.abs(Integer.parseInt(columnValue)) + " days ago");
+                    textView.setTextColor(Color.parseColor("#b03428"));
+                }
+                textView.setTypeface(null, Typeface.BOLD);
+            }
+
+            textView.setPadding(10, 10, 10, 10);
+            if (contractColumn.getKey().equals("contract_id")) {
+                contractRow.addView(btn);
+            } else {
+                contractRow.addView(textView);
+            }
         }
 
         table.addView(contractRow);
