+++ app/src/main/java/app/agrivest/android/utils/SQLiteHelper.java
@@ -1,4 +1,4 @@
-package app.agrivest.android;
+package app.agrivest.android.utils;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
