+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingTop="50dp"
     android:layout_margin="10dp"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <ImageView
         android:layout_width="match_parent"
