+++ app/src/main/res/layout/fragment_contract_details.xml
@@ -1,14 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <androidx.appcompat.widget.Toolbar
-    android:id="@+id/toolbar"
-    android:layout_width="match_parent"
-    android:layout_height="?attr/actionBarSize"
-    android:background="@android:color/white" />
+    android:layout_height="match_parent">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -120,11 +113,11 @@
                     android:layout_height="wrap_content"
                     android:layout_gravity="left"
                     android:layout_weight="1"
-                    android:text="Chassis Number"
+                    android:text="Model"
                     android:textSize="15sp" />
 
                 <TextView
-                    android:id="@+id/chassis_number_TV"
+                    android:id="@+id/model_TV"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_gravity="right"
@@ -183,8 +176,39 @@
                     android:text="Total Payable"
                     android:textSize="15sp" />
 
-                <TextView
+                <com.google.android.material.button.MaterialButton
                     android:id="@+id/total_payable_TV"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="right"
+                    android:layout_weight="1"
+                    style="?attr/materialButtonOutlinedStyle"
+                    android:textSize="15sp" />
+
+            </LinearLayout>
+
+            <View
+                android:layout_width="match_parent"
+                android:layout_height="1dp"
+                android:layout_marginTop="5dp"
+                android:layout_marginBottom="5dp"
+                android:background="#c0c0c0" />
+
+            <LinearLayout
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <TextView
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="left"
+                    android:layout_weight="1"
+                    android:text="Chassis Number"
+                    android:textSize="15sp" />
+
+                <TextView
+                    android:id="@+id/chassis_number_TV"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_gravity="right"
@@ -200,50 +224,31 @@
                 android:layout_marginBottom="5dp"
                 android:background="#c0c0c0" />
 
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Payment Mode" />
-
-            <Spinner
-                android:id="@+id/payment_method_SP"
+            <LinearLayout
                 android:layout_width="fill_parent"
-                android:drawSelectorOnTop="true"
-                android:entries="@array/payment_types"
-                android:layout_height="wrap_content" />
-
-            <TextView
-                android:id="@+id/due_date_TV"
-                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="Due Date" />
+                android:orientation="horizontal">
 
-            <EditText
-                android:id="@+id/due_date_ET"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="Due Date" />
+                <TextView
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="left"
+                    android:layout_weight="1"
+                    android:text="Contact"
+                    android:textSize="15sp" />
 
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Amount" />
+                <TextView
+                    android:id="@+id/contact_TV"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="right"
+                    android:layout_weight="1"
+                    android:textSize="15sp" />
 
-            <EditText
-                android:id="@+id/amount_ET"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="349000"
-                android:inputType="number" />
-
-            <Button
-                android:id="@+id/issue_receipt_BT"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="20dp"
-                android:text="Issue Receipt" />
+            </LinearLayout>
 
         </LinearLayout>
+
     </ScrollView>
 
-</LinearLayout>
+</RelativeLayout>
