+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,7 @@
         <item>Cash</item>
         <item>Check</item>
     </string-array>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
